Rename vuex store module
