+++ src/components/CampsiteFilter/partials/PageSize.vue
@@ -11,11 +11,11 @@ import { mapGetters, mapState, mapActions } from "vuex";
 export default {
   name: "PageSize",
   computed: {
-    ...mapState(["searchResults", "authentication"]),
+    ...mapState(["data", "authentication"]),
     ...mapGetters(["i18n"]),
     limit: {
       get() {
-        return this.searchResults.limit;
+        return this.data.limit;
       },
       set(value) {
         this.applyPageLimit({ value, token: this.authentication.token });
