+++ src/store/index.ts
@@ -3,7 +3,7 @@ import Vuex from "vuex";
 
 import * as app from "./modules/app";
 import * as authentication from "./modules/authentication";
-import * as searchResults from "./modules/searchResults";
+import * as data from "./modules/data";
 import * as detailPage from "./modules/detailPage";
 import * as translations from "./modules/translations";
 import * as wishlist from "./modules/wishlist";
@@ -14,7 +14,7 @@ export default new Vuex.Store({
   modules: {
     app,
     authentication,
-    searchResults,
+    data,
     detailPage,
     wishlist,
     translations
